+++ src/components/molecules/LoginButton.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { Button, Icon } from "semantic-ui-react";
 import { useSelector } from "react-redux";
 
-import { useAuthState } from "../../hooks/useAuthState";
+import { useAuth } from "../../hooks/useAuth";
 
 export const LoginButton: React.FC = () => {
-  const { login } = useAuthState();
+  const { login } = useAuth();
   const { user } = useSelector((state) => state.auth);
   return (
     <Button color="black" size="big" onClick={() => login()} disabled={!!user}>
